Take into account the notif_timeout parameter from the config file. Change it's precision to 0.1 sec. (JIRA: CXX-2728)
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@51604 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netschedule/ns_queue_parameters.hpp
@@ -47,7 +47,7 @@ struct SQueueParameters
 
     /// General parameters, reconfigurable at run time
     int             timeout;
-    int             notif_timeout;
+    double          notif_timeout;
     int             run_timeout;
     string          program_name;
     int             failed_retries;
